init: re-open console before pivoting root
In configurations where the console driver is a boot module, like
virtio_console, this makes sure all further userspace output actually
appears on that console, instead of being sent to the void.

A device manager like mdev needs to run after the modules are loaded
to actually locate and populate the device nodes, so that /dev/console
is properly setup in userspace to direct to the console= argument.
That means we can only re-open the console after running nlplug-findfs:
at that point, there's so few possible messages left from our own init that
it's easier to just let busybox's switch_root applet handle it instead of
something like `exec </dev/console >/dev/console 2>&1`.